third_party/WebKit/LayoutTests/inspector-protocol/dom/dom-distributed-nodes.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/inspector-protocol-test.js"></script>
<script type="text/javascript" src="../../http/tests/inspector-protocol/dom-protocol-test.js"></script>
<script>

function createShadowRootFromTemplate(root, selector, templateId)
{
    var shadowHost = root.querySelector(selector);
    var shadowRoot = shadowHost.createShadowRoot();
    var template = document.querySelector(templateId);
    var clone = document.importNode(template.content, true);
    shadowRoot.appendChild(clone);
    return shadowHost;
}

function initOldestShadowRoot()
{
    createShadowRootFromTemplate(document, "#shadowHost", "#oldestShadowRootTemplate");
}

function initYoungestShadowRoot()
{
    createShadowRootFromTemplate(document, "#shadowHost", "#youngestShadowRootTemplate");
}

function initInnerShadowRoot()
{
    var shadowHost = document.querySelector("#shadowHost");
    var innerShadowHost = createShadowRootFromTemplate(shadowHost.shadowRoot, ".innerShadowHost", "#innerShadowRootTemplate");
    innerShadowHost.id = "innerShadowHost";
}

var lastDistributedNodeId = 0;

function addDistributedNode(oldest)
{
    var node = document.createElement("div");
    node.classList.add(oldest ? "distributeMeToOldest" : "distributeMeToYoungest");
    node.classList.add("distributeMeAsWell_" + (++lastDistributedNodeId));
    var shadowHost = document.querySelector("#shadowHost");
    shadowHost.appendChild(node);
}

function addDistributedNodeToOldest()
{
    addDistributedNode(true);
}

function addDistributedNodeToYoungest()
{
    addDistributedNode(false);
}

function removeDistributedNodes()
{
    // Removing inner shadow host first
    var shadowHost = document.querySelector("#shadowHost");
    var innerShadowHost = shadowHost.shadowRoot.querySelector(".innerShadowHost");
    innerShadowHost.remove();

    var shadowHost = document.querySelector("#shadowHost");
    shadowHost.textContent = "";
}

function putTextIntoContentNode()
{
    var contentNode = document.querySelector("body /deep/ content[select='.distributeMeToYoungest']");
    contentNode.textContent = "fallback text";
}

function removeContentChildren()
{
    var contentNode = document.querySelector("body /deep/ content[select='.distributeMeToYoungest']");
    contentNode.textContent = "";
}

function test()
{
    var documentNodeId;
    var contentNodeId;
    var shadowHostNodeId;
    var nodeInfo = {};
    InspectorTest.trackGetChildNodesEvents(nodeInfo);
    InspectorTest.eventHandler["DOM.childNodeInserted"] = childNodeInserted;

    function childNodeInserted(message)
    {
        InspectorTest.addNode(nodeInfo, message.params.node);
    }

    InspectorTest.eventHandler["DOM.shadowHostDistributionInvalidated"] = shadowHostDistributionInvalidated;

    InspectorTest.evaluateInInspectedPage("initOldestShadowRoot()", onOldestShadowRootInitialized);

    function onOldestShadowRootInitialized()
    {
        InspectorTest.requestDocumentNodeId(onDocumentNodeReceived);
    }

    function onDocumentNodeReceived(nodeId)
    {
        documentNodeId = nodeId;
        InspectorTest.log("Loading content node...");
        InspectorTest.requestNodeId(documentNodeId, "body /deep/ content", onContentNodeReceived);
    }

    function onContentNodeReceived(nodeId)
    {
        contentNodeId = nodeId;

        InspectorTest.log("Loading shadow host node...");
        InspectorTest.requestNodeId(documentNodeId, "#shadowHost", onShadowHostNodeReceived);
    }

    function onShadowHostNodeReceived(nodeId)
    {
        shadowHostNodeId = nodeId;

        InspectorTest.log("Adding distributed node before distributed nodes requested...");
        InspectorTest.evaluateInInspectedPage("addDistributedNodeToOldest()", evaluated);

        function distributedNodeAdded()
        {
            // Make one more round trip to make sure we did not receive any unexpected invalidation events.
            InspectorTest.evaluateInInspectedPage("var dummy = 42;", evaluated);
        }

        function evaluated()
        {
            requestDistributedNodesForNodes([shadowHostNodeId], onDistributionRequested);
        }
    }

    function onDistributionRequested()
    {
        InspectorTest.log("Adding distributed node...");
        InspectorTest.evaluateInInspectedPage("addDistributedNodeToOldest()");
        waitForShadowHostDistributionInvalidatedAndReloaded(distributedNodeChangedAfterFirstAdding);
    }

    function distributedNodeChangedAfterFirstAdding()
    {
        InspectorTest.log("Shadow host distributed nodes request finished.");

        InspectorTest.log("Adding another node to be distributed...");
        InspectorTest.evaluateInInspectedPage("addDistributedNodeToOldest()");
        waitForShadowHostDistributionInvalidatedAndReloaded(distributedNodeChangedAfterSecondAdding);
    }

    function distributedNodeChangedAfterSecondAdding()
    {
        InspectorTest.log("Adding youngest shadow root...");
        InspectorTest.evaluateInInspectedPage("initYoungestShadowRoot()");
        waitForShadowHostDistributionInvalidatedAndReloaded(distributedNodeChangedAfterAddingYoungestShadowRoot);
    }

    function distributedNodeChangedAfterAddingYoungestShadowRoot()
    {
        InspectorTest.log("Adding inner shadow root node...");
        InspectorTest.evaluateInInspectedPage("initInnerShadowRoot()");
        InspectorTest.requestNodeId(documentNodeId, "body /deep/ #innerShadowHost", onInnerShadowHostNodeReceived);
    }

    function onInnerShadowHostNodeReceived(innerShadowHostId)
    {
        requestDistributedNodesForNodes([innerShadowHostId], distributedNodeChangedAfterAddingInnerShadowRoot);
    }

    function distributedNodeChangedAfterAddingInnerShadowRoot()
    {
        InspectorTest.log("Inner shadow host distributed nodes request finished.");

        InspectorTest.log("Removing shadow host children...");
        InspectorTest.evaluateInInspectedPage("removeDistributedNodes()");
        waitForShadowHostDistributionInvalidatedAndReloaded(distributedNodeChangedAfterClearingShadowHost);
    }

    function distributedNodeChangedAfterClearingShadowHost()
    {
        // Here we make sure that we work correctly for text nodes as well.
        InspectorTest.log("Putting text into content node...");
        InspectorTest.evaluateInInspectedPage("putTextIntoContentNode()");
        waitForShadowHostDistributionInvalidatedAndReloaded(distributedNodeChangedAfterPuttingTextIntoContextNode);
    }

    function distributedNodeChangedAfterPuttingTextIntoContextNode()
    {
        InspectorTest.log("Removing content node children...");
        InspectorTest.evaluateInInspectedPage("removeContentChildren()");
        waitForShadowHostDistributionInvalidatedAndReloaded(distributedNodeChangedAfterClearingContentNode);
    }

    function distributedNodeChangedAfterClearingContentNode()
    {
        InspectorTest.log("SUCCESS");
        InspectorTest.completeTest();
    }

    function nodeToString(nodeId)
    {
        var node = nodeInfo[nodeId];
        if (node.shadowRootType)
            return "#shadowRoot";
        if (node.nodeName === "#text")
            return "#text(" + node.nodeValue.replace(/\n/, "\\n") + ")";
        var attributeString = "";
        for (var i = 0; node.attributes && i < node.attributes.length; ++i) {
            if (i % 2 === 0)
                continue;
            attributeString += " " + node.attributes[i - 1] + (node.attributes[i] ? "=" + node.attributes[i] : "");
        }
        var nodeString = node.localName + attributeString;
        return "<" + nodeString + ">";
    }

    var nextStep;
    function waitForShadowHostDistributionInvalidatedAndReloaded(next)
    {
        nextStep = next;
    }

    function shadowHostDistributionInvalidated(message)
    {
        var shadowHostIds = message.params.nodeIds;
        InspectorTest.log("Shadow host distribution invalidated");

        var nodeIds = [];
        for (var i = 0; i < shadowHostIds.length; ++i)
            nodeIds.push(shadowHostIds[i]);
        requestDistributedNodesForNodes(nodeIds, nextStep);
        nextStep = null;
    }

    function requestDistributedNodesForNodes(nodeIds, callback)
    {
        InspectorTest.log("Requesting distributed nodes for: ");
        for (var nodeId of nodeIds) {
            InspectorTest.log("  - " + nodeToString(nodeId));
            requestShadowHostDistribution(nodeId, distributedNodesLoaded);
        }

        var callbacksLeft = nodeIds.length;
        function distributedNodesLoaded()
        {
            if (--callbacksLeft)
                return;
            if (callback)
                callback();
        }
    }

    function requestShadowHostDistribution(shadowHostId, callback)
    {
        InspectorTest.sendCommandOrDie("DOM.requestShadowHostDistributedNodes", { "nodeId": shadowHostId }, distributionLoaded);

        function distributionLoaded(result)
        {
            var insertionPointDistributions = result.insertionPointDistributions;

            InspectorTest.log("  Shadow host distribution for " + nodeToString(shadowHostId) + ":");
            for (var insertionPointDistribution of insertionPointDistributions) {
                var insertionPointId = insertionPointDistribution.nodeId;
                var distributedNodes = insertionPointDistribution.distributedNodes;
                var distributedNodeIds = insertionPointDistribution.distributedNodeIds || [];

                InspectorTest.log("    Distributed nodes set for " + nodeToString(insertionPointId) + " :");
                for (var distributedNode of distributedNodes) {
                    var distributedNodeId = distributedNode.nodeId;

                    var destinationInsertionPointIds = distributedNode.destinationInsertionPointIds;
                    var throughString = "";
                    if (destinationInsertionPointIds)
                        throughString = " (through " + destinationInsertionPointIds.map(nodeToString).join(", ") + ")";
                    InspectorTest.log("     - " + nodeToString(distributedNodeId) + throughString);
                }
            }
            callback();
        }
    }
}
</script>
</head>
<body onload="runTest()">
<p>Tests that distributed nodes are correctly reported in the protocol as long as their updates.</p>
<template id="youngestShadowRootTemplate">
    <div class="youngestShadowMain">
        <shadow></shadow>
        <content select=".distributeMeToYoungest"><div id="fallbackYoungest"></div></content>
        <div class="innerShadowHost">
            <content in-youngest-shadow-root select=".distributeMeToInner"></content>
        </div>
    </div>
</template>
<template id="oldestShadowRootTemplate">
    <div class="oldestShadowMain">
        <content select=".distributeMeToOldest"><div id="fallbackOldest"></div></content>
    </div>
</template>
<template id="innerShadowRootTemplate">
    <div class="innerShadowMain">
        <content in-inner-shadow-root select=".distributeMeToInner"></content>
    </div>
</template>
<div id="shadowHost">
    <div class="distributeMeToYoungest original">
        youngest distributed text
    </div>
    <div class="distributeMeToOldest original">
        oldest distributed text
    </div>
    <div class="distributeMeToInner original">
        oldest distributed text
    </div>
    <div class="distributeMeToInner original2">
        oldest distributed text
    </div>
</div>
</body>
</html>
