third_party/WebKit/LayoutTests/inspector-protocol/dom/dom-distributed-nodes-expected.txt

Tests that distributed nodes are correctly reported in the protocol as long as their updates.

Loading content node...
Loading shadow host node...
Adding distributed node before distributed nodes requested...
Requesting distributed nodes for: 
  - <div id=shadowHost>
  Shadow host distribution for <div id=shadowHost>:
    Distributed nodes set for <content select=.distributeMeToOldest> :
     - <div class=distributeMeToOldest original>
     - <div class=distributeMeToOldest distributeMeAsWell_1>
Adding distributed node...
Shadow host distribution invalidated
Requesting distributed nodes for: 
  - <div id=shadowHost>
  Shadow host distribution for <div id=shadowHost>:
    Distributed nodes set for <content select=.distributeMeToOldest> :
     - <div class=distributeMeToOldest original>
     - <div class=distributeMeToOldest distributeMeAsWell_1>
     - <div class=distributeMeToOldest distributeMeAsWell_2>
Shadow host distributed nodes request finished.
Adding another node to be distributed...
Shadow host distribution invalidated
Requesting distributed nodes for: 
  - <div id=shadowHost>
  Shadow host distribution for <div id=shadowHost>:
    Distributed nodes set for <content select=.distributeMeToOldest> :
     - <div class=distributeMeToOldest original>
     - <div class=distributeMeToOldest distributeMeAsWell_1>
     - <div class=distributeMeToOldest distributeMeAsWell_2>
     - <div class=distributeMeToOldest distributeMeAsWell_3>
Adding youngest shadow root...
Shadow host distribution invalidated
Requesting distributed nodes for: 
  - <div id=shadowHost>
  Shadow host distribution for <div id=shadowHost>:
    Distributed nodes set for <shadow> :
     - <div class=oldestShadowMain>
    Distributed nodes set for <content select=.distributeMeToYoungest> :
     - <div class=distributeMeToYoungest original>
    Distributed nodes set for <content in-youngest-shadow-root select=.distributeMeToInner> :
     - <div class=distributeMeToInner original>
     - <div class=distributeMeToInner original2>
    Distributed nodes set for <content select=.distributeMeToOldest> :
     - <div class=distributeMeToOldest original>
     - <div class=distributeMeToOldest distributeMeAsWell_1>
     - <div class=distributeMeToOldest distributeMeAsWell_2>
     - <div class=distributeMeToOldest distributeMeAsWell_3>
Adding inner shadow root node...
Requesting distributed nodes for: 
  - <div class=innerShadowHost>
  Shadow host distribution for <div class=innerShadowHost>:
    Distributed nodes set for <content in-inner-shadow-root select=.distributeMeToInner> :
     - <div class=distributeMeToInner original> (through <content in-youngest-shadow-root select=.distributeMeToInner>, <content in-inner-shadow-root select=.distributeMeToInner>)
     - <div class=distributeMeToInner original2> (through <content in-youngest-shadow-root select=.distributeMeToInner>, <content in-inner-shadow-root select=.distributeMeToInner>)
Inner shadow host distributed nodes request finished.
Removing shadow host children...
Shadow host distribution invalidated
Requesting distributed nodes for: 
  - <div id=shadowHost>
  Shadow host distribution for <div id=shadowHost>:
    Distributed nodes set for <shadow> :
     - <div class=oldestShadowMain>
    Distributed nodes set for <content select=.distributeMeToYoungest> :
     - <div id=fallbackYoungest>
    Distributed nodes set for <content select=.distributeMeToOldest> :
     - <div id=fallbackOldest>
Putting text into content node...
Shadow host distribution invalidated
Requesting distributed nodes for: 
  - <div id=shadowHost>
  Shadow host distribution for <div id=shadowHost>:
    Distributed nodes set for <shadow> :
     - <div class=oldestShadowMain>
    Distributed nodes set for <content select=.distributeMeToYoungest> :
     - #text(fallback text)
    Distributed nodes set for <content select=.distributeMeToOldest> :
     - <div id=fallbackOldest>
Removing content node children...
Shadow host distribution invalidated
Requesting distributed nodes for: 
  - <div id=shadowHost>
  Shadow host distribution for <div id=shadowHost>:
    Distributed nodes set for <shadow> :
     - <div class=oldestShadowMain>
    Distributed nodes set for <content select=.distributeMeToYoungest> :
    Distributed nodes set for <content select=.distributeMeToOldest> :
     - <div id=fallbackOldest>
SUCCESS
